Move source of data off of Airtable
Because:

Airtable's API changes for images mean I have to download the images for
every build instead of relying on their URLs to proxy, which is wasteful
and makes for much longer builds.

Solution:

Store the images elsewhere, and replace the metadata in Airtable with
JSON stored alongside the images.

- Put all the originals, pre-built thumbnails, and JSON on a WebDAV
  server I control
- Fetch the JSON on build, creating Netlfiy proxying redirects to keep
  the local paths reasonable.
- Remove all the Airtable code
- Remove all the code copying the source files from Airtable
- Start working on script to upload *new* images to the WebDav server